// extensions/clickup-tasks/src/types/clickup.ts
interface ClickUpStatus {
  color: string;
  id?: string;
  orderindex?: number;
  status: string;
  type?: string;
}

export interface ClickUpUser {
  color?: string;
  email?: string;
  id: number;
  initials?: string;
  profilePicture?: string | null;
  username: string;
}

export interface ClickUpPriority {
  color: string;
  id: string;
  orderindex?: string;
  priority: string;
}

interface ClickUpTag {
  creator?: number;
  name: string;
  tag_bg: string;
  tag_fg: string;
}

interface ClickUpCustomField {
  date_created?: string;
  hide_from_guests?: boolean;
  id: string;
  name: string;
  required?: boolean;
  type: string;
  type_config?: Record<string, unknown>;
  value?: string | number | boolean | null;
}

export interface ClickUpList {
  access?: boolean;
  archived?: boolean;
  content?: string;
  due_date?: string | null;
  folder?: ClickUpFolder;
  id: string;
  name: string;
  priority?: ClickUpPriority | null;
  space?: ClickUpSpace;
  start_date?: string | null;
  statuses?: ClickUpStatus[];
  task_count?: number;
  team_id?: string;
  url?: string;
}

export interface ClickUpFolder {
  access?: boolean;
  hidden?: boolean;
  id: string;
  lists?: ClickUpList[];
  name: string;
  space?: ClickUpSpace;
  task_count?: number;
}

export interface ClickUpSpace {
  access?: boolean;
  archived?: boolean;
  features?: Record<string, unknown>;
  id: string;
  multiple_assignees?: boolean;
  name: string;
  private?: boolean;
  statuses?: ClickUpStatus[];
}